finally got it debugged had to drop it into a second instance of the idea and copy files over to get it to run through the debugger.

+++ phone_source/phone.c
@@ -24,9 +24,10 @@ int main(){
         }
     }
     else{
-         phoneNumber[strlen(phoneNumber)-1] = '\0';
+        phoneNumber[strlen(phoneNumber)-1] = '\0';
     }
-    len = strlen(phoneNumber);
+
+    len = (int)strlen(phoneNumber);
     formatNumber(len, phoneNumber);
